server/requirements.txt
# ftp获取 ocr识别
pyftpdlib
oss2
requests
easyocr
toml
pytest
reloadium
pillow
pillow-avif-plugin
paddleocr
opencv-python
opencv-python-headless
numpy
easyocr
ultralytics
paddlepaddle
requests


#  处理数据入库
python-dotenv
pymongo
fastapi[all]
pydantic
passlib
jinja2
fastapi-mail
fastapi-mail[aioredis]
fastapi-mail[httpx]
passlib[bcrypt]
#fastapi-jwt-auth
#fastapi-jwt-auth[asymmetric]
uvicorn
pydantic-settings
python-jose[cryptography]
sqlalchemy
oss2
python-multipart
python-dotenv
Pillow
pillow-avif-plugin